requirements.txt
# Azure Key Vault
azure-identity>=1.12.0
azure-keyvault-secrets>=4.8.0

# Data handling
pandas>=1.3.0
numpy>=1.21.0
openpyxl>=3.0.0
pymysql>=1.0.2

# Embeddings
sentence-transformers>=2.2.0
torch>=1.9.0
tqdm>=4.59.0

# Google Drive / Sheets access
gspread>=5.0.0
google-auth>=2.0.0
google-auth-oauthlib>=0.5.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0

# Visualization
matplotlib>=3.4.0
scipy>=1.7.0
plotly>=5.0.0

# Azure Storage & Database
azure-storage-blob>=12.0.0
sqlalchemy>=1.4.0
pyodbc>=4.0.0

# Configuration
python-dotenv>=0.19.0
